# Improve performance on Linux with many network interfaces