@willscott 2025-08-14
It would be great if we could figure out 3 years later how to upstream this into the standard library, rather than needing this utility layer for each client of calls to `net.Interfaces`. I wonder if it's worth another round of pinging maintainers on a resolution path here?